Fixing. Files: d 3-print_all.c 3-main.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -31,7 +31,7 @@ void print_all(const char * const format, ...)
 				printf("%d", i);
 				break;
 			case 'f':
-                                f = (float) va_arg(args, double);
+                                f = va_arg(args, double);
                                 printf("%f", f);
                                 break;
                         case 's':
@@ -45,6 +45,7 @@ void print_all(const char * const format, ...)
 					printf("%s", s);
 				}
                                 break;
+
 			default:
 				break;
 		}
